CLI: Print out the new operation-related events logged through the bridge to the standard output
