+++ cli/crates/cli/src/dev.rs
@@ -53,6 +53,16 @@ pub fn dev(search: bool, watch: bool, external_port: u16, tracing: bool) -> Resu
                 } => {
                     report::udf_message(udf_kind, &udf_name, &message, level);
                 }
+                ServerMessage::OperationStarted { request_id, name } => {
+                    report::operation_started(request_id, name);
+                }
+                ServerMessage::OperationCompleted {
+                    request_id,
+                    name,
+                    duration,
+                } => {
+                    report::operation_completed(request_id, name, duration);
+                }
                 ServerMessage::CompilationError(error) => report::error(&CliError::CompilationError(error)),
             }
         }
