+++ cli/crates/cli/src/output/report.rs
@@ -149,6 +149,14 @@ pub fn udf_message(udf_kind: UdfKind, udf_name: &str, message: &str, level: UdfM
     }
 }
 
+pub fn operation_started(_request_id: String, _name: Option<String>) {}
+
+pub fn operation_completed(_request_id: String, name: Option<String>, duration: std::time::Duration) {
+    // FIXME: Add operation type.
+    let formatted_name = name.map(|name| format!(" {name}")).unwrap_or_default();
+    println!("→{formatted_name}  {}ms", duration.as_millis());
+}
+
 pub fn reload<P: AsRef<Path>>(path: P) {
     println!(
         "🔄 Detected a change in {path}, reloading",
