+++ cli/crates/server/Cargo.toml
@@ -34,6 +34,7 @@ reqwest = { version = "0.11", features = [
 ], default-features = false }
 serde = { version = "1", features = ["serde_derive"] }
 serde_json = "1"
+serde_with = "3"
 slug = "0.1"
 sqlx = { version = "0.6", features = [
   "runtime-tokio-rustls",
