+++ cli/crates/server/src/bridge/mod.rs
@@ -1,6 +1,7 @@
 mod api_counterfeit;
 mod consts;
 mod listener;
+mod log;
 mod search;
 mod server;
 mod sqlite;
