+++ cli/crates/server/src/bridge/server.rs
@@ -1,52 +1,38 @@
-use super::api_counterfeit::registry::Registry;
-use super::api_counterfeit::search::{QueryExecutionRequest, QueryExecutionResponse};
 use super::consts::{DATABASE_FILE, DATABASE_URL_PREFIX, PREPARE};
-use super::search::Index;
-use super::types::{Mutation, Operation, Record, UdfInvocation};
-use crate::bridge::api_counterfeit::registry::VersionedRegistry;
+use super::types::{Mutation, Operation, Record};
+use super::udf::UdfBuild;
 use crate::bridge::errors::ApiError;
 use crate::bridge::listener;
+use crate::bridge::log::log_event_endpoint;
+use crate::bridge::search::search_endpoint;
 use crate::bridge::types::{Constraint, ConstraintKind, OperationKind};
+use crate::bridge::udf::invoke_udf_endpoint;
 use crate::errors::ServerError;
 use crate::event::{wait_for_event, Event};
 use crate::types::ServerMessage;
 use axum::extract::State;
 use axum::Json;
 use axum::{http::StatusCode, routing::post, Router};
-use common::environment::{Environment, Project};
+use common::environment::Project;
 
 use common::types::UdfKind;
 use sqlx::query::{Query, QueryAs};
 use sqlx::{migrate::MigrateDatabase, query, query_as, sqlite::SqlitePoolOptions, Sqlite, SqlitePool};
 use tokio::fs;
-use tokio::sync::{Mutex, Notify};
+use tokio::sync::Mutex;
 
 use std::collections::HashMap;
-use std::fs::File;
-use std::io::BufReader;
 use std::net::{Ipv4Addr, SocketAddr};
 use std::sync::Arc;
 
 use tower_http::trace::TraceLayer;
 
-pub enum UdfBuild {
-    InProgress {
-        notify: Arc<Notify>,
-    },
-    Succeeded {
-        #[allow(dead_code)]
-        miniflare_handle: tokio::task::JoinHandle<()>,
-        worker_port: u16,
-    },
-    Failed,
-}
-
-struct HandlerState {
-    pool: SqlitePool,
-    bridge_sender: tokio::sync::mpsc::Sender<ServerMessage>,
-    udf_builds: Mutex<std::collections::HashMap<(String, UdfKind), UdfBuild>>,
-    environment_variables: HashMap<String, String>,
-    tracing: bool,
+pub struct HandlerState {
+    pub pool: SqlitePool,
+    pub bridge_sender: tokio::sync::mpsc::Sender<ServerMessage>,
+    pub udf_builds: Mutex<std::collections::HashMap<(String, UdfKind), UdfBuild>>,
+    pub environment_variables: HashMap<String, String>,
+    pub tracing: bool,
 }
 
 async fn query_endpoint(
@@ -112,157 +98,6 @@ async fn mutation_endpoint(
     Ok(StatusCode::OK)
 }
 
-async fn search_endpoint(
-    State(handler_state): State<Arc<HandlerState>>,
-    Json(request): Json<QueryExecutionRequest>,
-) -> Result<Json<QueryExecutionResponse>, ApiError> {
-    let project = Project::get();
-
-    let registry: Registry = {
-        let path = &project.registry_path;
-        let file = File::open(path).map_err(|err| {
-            error!("Failed to open {path:?}: {err:?}");
-            ApiError::ServerError
-        })?;
-        let reader = BufReader::new(file);
-        let versioned: VersionedRegistry = serde_json::from_reader(reader).map_err(|err| {
-            error!("Failed to deserialize registry: {err:?}");
-            ApiError::ServerError
-        })?;
-        versioned.registry
-    };
-
-    let response = Index::build(&handler_state.pool, &request.entity_type, &registry.search_config)
-        .await?
-        .search(request.query, request.pagination)?;
-    Ok(Json(response))
-}
-
-#[allow(clippy::too_many_lines)]
-async fn invoke_udf_endpoint(
-    State(handler_state): State<Arc<HandlerState>>,
-    Json(payload): Json<UdfInvocation>,
-) -> Result<Json<serde_json::Value>, ApiError> {
-    use futures_util::TryFutureExt;
-
-    trace!("UDF invocation\n\n{:#?}\n", payload);
-
-    let environment = Environment::get();
-    let udf_kind = payload.udf_kind;
-
-    let udf_worker_port = loop {
-        let notify = {
-            let mut udf_builds: tokio::sync::MutexGuard<'_, _> = handler_state.udf_builds.lock().await;
-
-            if let Some(udf_build) = udf_builds.get(&(payload.name.clone(), udf_kind)) {
-                match udf_build {
-                    UdfBuild::Succeeded { worker_port, .. } => break *worker_port,
-                    UdfBuild::Failed => return Err(ApiError::UdfInvocation),
-                    UdfBuild::InProgress { notify } => {
-                        // If the resolver build happening within another invocation has been cancelled
-                        // due to the invocation having been interrupted by the HTTP client, start a new build.
-                        if Arc::strong_count(notify) == 1 {
-                            notify.clone()
-                        } else {
-                            let notify = notify.clone();
-                            drop(udf_builds);
-                            notify.notified().await;
-                            continue;
-                        }
-                    }
-                }
-            } else {
-                let notify = Arc::new(Notify::new());
-                udf_builds.insert(
-                    (payload.name.clone(), udf_kind),
-                    UdfBuild::InProgress { notify: notify.clone() },
-                );
-                notify
-            }
-        };
-
-        let start = std::time::Instant::now();
-        handler_state
-            .bridge_sender
-            .send(ServerMessage::StartUdfBuild {
-                udf_kind,
-                udf_name: payload.name.clone(),
-            })
-            .await
-            .unwrap();
-
-        let tracing = handler_state.tracing;
-
-        match crate::udf_builder::build(
-            environment,
-            &handler_state.environment_variables,
-            udf_kind,
-            &payload.name,
-            tracing,
-        )
-        .and_then(|(package_json_path, wrangler_toml_path)| {
-            super::udf::spawn_miniflare(udf_kind, &payload.name, package_json_path, wrangler_toml_path, tracing)
-        })
-        .await
-        {
-            Ok((miniflare_handle, worker_port)) => {
-                handler_state.udf_builds.lock().await.insert(
-                    (payload.name.clone(), udf_kind),
-                    UdfBuild::Succeeded {
-                        miniflare_handle,
-                        worker_port,
-                    },
-                );
-                notify.notify_waiters();
-
-                handler_state
-                    .bridge_sender
-                    .send(ServerMessage::CompleteUdfBuild {
-                        udf_kind,
-                        udf_name: payload.name.clone(),
-                        duration: start.elapsed(),
-                    })
-                    .await
-                    .unwrap();
-
-                break worker_port;
-            }
-            Err(err) => {
-                error!(
-                    "Build of {udf_kind} '{udf_name}' failed: {err:?}",
-                    udf_name = payload.name
-                );
-                handler_state
-                    .bridge_sender
-                    .send(ServerMessage::CompilationError(format!(
-                        "{udf_kind} '{udf_name}' failed to build: {err}",
-                        udf_name = payload.name
-                    )))
-                    .await
-                    .unwrap();
-            }
-        };
-
-        handler_state
-            .udf_builds
-            .lock()
-            .await
-            .insert((payload.name.clone(), udf_kind), UdfBuild::Failed);
-        notify.notify_waiters();
-        return Err(ApiError::UdfInvocation);
-    };
-
-    super::udf::invoke(
-        &handler_state.bridge_sender,
-        udf_worker_port,
-        udf_kind,
-        &payload.name,
-        &payload.payload,
-    )
-    .await
-    .map(Json)
-}
-
 pub async fn start(
     port: u16,
     worker_port: u16,
@@ -312,8 +147,8 @@ pub async fn start(
         .route("/query", post(query_endpoint))
         .route("/mutation", post(mutation_endpoint))
         .route("/search", post(search_endpoint))
-        .route("/invoke-resolver", post(invoke_udf_endpoint)) // FIXME: remove after API repo is switched.
         .route("/invoke-udf", post(invoke_udf_endpoint))
+        .route("/log-event", post(log_event_endpoint))
         .with_state(handler_state.clone())
         .layer(TraceLayer::new_for_http());
