+++ cli/crates/server/src/bridge/types.rs
@@ -70,9 +70,26 @@ pub struct RecordDocument {
 
 #[derive(Deserialize, Debug)]
 pub struct UdfInvocation {
-    #[serde(alias = "resolver_name")] // FIXME: remove after api repo is updated
     pub name: String,
     pub payload: serde_json::Value,
-    #[serde(default)] // FIXME: remove after api repo is updated
     pub udf_kind: UdfKind,
+}
+
+#[serde_with::serde_as]
+#[derive(Deserialize, Debug)]
+pub enum LogEventType {
+    OperationStarted {
+        name: Option<String>,
+    },
+    OperationCompleted {
+        name: Option<String>,
+        #[serde_as(as = "serde_with::DurationMilliSeconds<u64>")]
+        duration: std::time::Duration,
+    },
+}
+
+#[derive(Deserialize, Debug)]
+pub struct LogEvent {
+    pub request_id: String,
+    pub r#type: LogEventType,
 }
