+++ cli/crates/server/src/bridge/udf.rs
@@ -1,14 +1,20 @@
 use std::process::Stdio;
+use std::sync::Arc;
 
 use crate::errors::UdfBuildError;
 use crate::types::ServerMessage;
 
+use axum::extract::State;
+use axum::Json;
 use common::types::UdfKind;
 use common::{environment::Environment, types::UdfMessageLevel};
 use futures_util::{pin_mut, TryFutureExt, TryStreamExt};
 use tokio::process::Command;
+use tokio::sync::Notify;
 
 use super::errors::ApiError;
+use super::server::HandlerState;
+use super::types::UdfInvocation;
 
 #[derive(serde::Deserialize)]
 #[serde(rename_all = "camelCase")]
@@ -23,6 +29,141 @@ struct UdfResponse {
     log_entries: Vec<UdfMessage>,
     #[serde(flatten)]
     rest: serde_json::Value,
+}
+
+pub enum UdfBuild {
+    InProgress {
+        notify: Arc<Notify>,
+    },
+    Succeeded {
+        #[allow(dead_code)]
+        miniflare_handle: tokio::task::JoinHandle<()>,
+        worker_port: u16,
+    },
+    Failed,
+}
+
+#[allow(clippy::too_many_lines)]
+pub async fn invoke_udf_endpoint(
+    State(handler_state): State<Arc<HandlerState>>,
+    Json(payload): Json<UdfInvocation>,
+) -> Result<Json<serde_json::Value>, ApiError> {
+    trace!("UDF invocation\n\n{:#?}\n", payload);
+
+    let environment = Environment::get();
+    let udf_kind = payload.udf_kind;
+
+    let udf_worker_port = loop {
+        let notify = {
+            let mut udf_builds: tokio::sync::MutexGuard<'_, _> = handler_state.udf_builds.lock().await;
+
+            if let Some(udf_build) = udf_builds.get(&(payload.name.clone(), udf_kind)) {
+                match udf_build {
+                    UdfBuild::Succeeded { worker_port, .. } => break *worker_port,
+                    UdfBuild::Failed => return Err(ApiError::UdfInvocation),
+                    UdfBuild::InProgress { notify } => {
+                        // If the resolver build happening within another invocation has been cancelled
+                        // due to the invocation having been interrupted by the HTTP client, start a new build.
+                        if Arc::strong_count(notify) == 1 {
+                            notify.clone()
+                        } else {
+                            let notify = notify.clone();
+                            drop(udf_builds);
+                            notify.notified().await;
+                            continue;
+                        }
+                    }
+                }
+            } else {
+                let notify = Arc::new(Notify::new());
+                udf_builds.insert(
+                    (payload.name.clone(), udf_kind),
+                    UdfBuild::InProgress { notify: notify.clone() },
+                );
+                notify
+            }
+        };
+
+        let start = std::time::Instant::now();
+        handler_state
+            .bridge_sender
+            .send(ServerMessage::StartUdfBuild {
+                udf_kind,
+                udf_name: payload.name.clone(),
+            })
+            .await
+            .unwrap();
+
+        let tracing = handler_state.tracing;
+
+        match crate::udf_builder::build(
+            environment,
+            &handler_state.environment_variables,
+            udf_kind,
+            &payload.name,
+            tracing,
+        )
+        .and_then(|(package_json_path, wrangler_toml_path)| {
+            super::udf::spawn_miniflare(udf_kind, &payload.name, package_json_path, wrangler_toml_path, tracing)
+        })
+        .await
+        {
+            Ok((miniflare_handle, worker_port)) => {
+                handler_state.udf_builds.lock().await.insert(
+                    (payload.name.clone(), udf_kind),
+                    UdfBuild::Succeeded {
+                        miniflare_handle,
+                        worker_port,
+                    },
+                );
+                notify.notify_waiters();
+
+                handler_state
+                    .bridge_sender
+                    .send(ServerMessage::CompleteUdfBuild {
+                        udf_kind,
+                        udf_name: payload.name.clone(),
+                        duration: start.elapsed(),
+                    })
+                    .await
+                    .unwrap();
+
+                break worker_port;
+            }
+            Err(err) => {
+                error!(
+                    "Build of {udf_kind} '{udf_name}' failed: {err:?}",
+                    udf_name = payload.name
+                );
+                handler_state
+                    .bridge_sender
+                    .send(ServerMessage::CompilationError(format!(
+                        "{udf_kind} '{udf_name}' failed to build: {err}",
+                        udf_name = payload.name
+                    )))
+                    .await
+                    .unwrap();
+            }
+        };
+
+        handler_state
+            .udf_builds
+            .lock()
+            .await
+            .insert((payload.name.clone(), udf_kind), UdfBuild::Failed);
+        notify.notify_waiters();
+        return Err(ApiError::UdfInvocation);
+    };
+
+    super::udf::invoke(
+        &handler_state.bridge_sender,
+        udf_worker_port,
+        udf_kind,
+        &payload.name,
+        &payload.payload,
+    )
+    .await
+    .map(Json)
 }
 
 async fn wait_until_udf_ready(worker_port: u16, udf_kind: UdfKind, udf_name: &str) -> Result<bool, reqwest::Error> {
